requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.3.0
crispy-bootstrap5==0.7
cx-Oracle==8.3.0
Django==4.2.5
django-bootstrap-v5==1.0.11
django-crispy-forms==2.0
django-filter==23.3
django-polymorphic==3.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
idna==3.4
marshmallow==3.17.0
packaging==23.2
Pillow==10.0.0
PyJWT==2.8.0
pytz==2023.3.post1
requests==2.31.0
soupsieve==2.5
sqlparse==0.4.4
transbank-sdk==4.0.0
tzdata==2023.3
urllib3==2.0.6
